requirements.txt
# Web框架
Flask==3.0.2
Flask-Session==0.7.0
Flask-Cors==4.0.0
python-dotenv==1.0.1

# LangChain相关
langchain-core==0.1.27
langchain-openai==0.0.8
langchain-ollama==0.0.5
langgraph==0.0.27

# OpenAI
openai==1.12.0

# 工具库
pydantic==2.6.3
typing-extensions==4.10.0

# 其他依赖
python-dateutil==2.8.2
requests==2.31.0
aiohttp==3.9.3

# 其他依赖
lightrag
graphviz>=0.20.1
gunicorn>=20.1.0  # 生产环境WSGI服务器 